requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.29.0
bcrypt==4.0.1
boto3==1.42.54
botocore==1.42.54
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.5.3
cryptography==46.0.5
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.23
fastavro==1.12.1
filelock==3.24.3
fsspec==2026.2.0
gotrue==2.12.4
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.3
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.7
packaging==26.0
passlib==1.7.4
pillow==12.1.1
postgrest==0.16.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.7.1
pydantic-settings==2.3.0
pydantic_core==2.18.2
Pygments==2.19.2
PyJWT==2.11.0
pypdf==6.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
realtime==1.0.6
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.4
rsa==4.9.1
s3transfer==0.16.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
storage3==0.7.7
StrEnum==0.4.15
supabase==2.4.6
supafunc==0.4.7
tokenizers==0.19.1
tqdm==4.67.3
typer==0.24.1
types-requests==2.32.4.20260107
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.30.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
